# apps/backend/openapi-quotes.yaml
openapi: 3.0.3
info:
  title: Pivotal Flow Quotes API
  description: |
    API for managing quotes in the Pivotal Flow system.
    
    ## Features
    - Create, read, update, and delete quotes
    - Quote status transitions (draft → pending → approved → sent → accepted)
    - Automatic calculations (subtotal, tax, discounts, total)
    - Multi-tenant support with organization isolation
    - Audit logging for all quote changes
    - Quote number generation with organization prefixes
    
    ## Authentication
    All endpoints require authentication via JWT token in the Authorization header.
    The token must include organization and user context for multi-tenant access.
    
    ## Multi-tenancy
    All quotes are scoped to the user's organization. Users can only access quotes
    within their own organization.
  version: 1.0.0
  contact:
    name: Pivotal Flow Team
    email: support@pivotalflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pivotalflow.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user and organization context.
        Token must include:
        - `sub`: User ID
        - `org`: Organization ID
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
  schemas:
    QuoteStatus:
      type: string
      enum: [draft, pending, approved, sent, accepted, rejected, cancelled]
      description: Current status of the quote
      example: "draft"
    QuoteType:
      type: string
      enum: [project, service, product, recurring, one_time]
      description: Type of quote
      example: "project"
    LineItemType:
      type: string
      enum: [service, product, material, travel, expense, discount, tax]
      description: Type of line item
      example: "service"
    DiscountType:
      type: string
      enum: [percentage, fixed_amount, per_unit]
      description: Type of discount
      example: "percentage"
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: string
          description: Amount as a decimal string for precision
          example: "100.00"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code
          example: "NZD"
    LineItem:
      type: object
      required:
        - lineNumber
        - description
        - quantity
        - unitPrice
      properties:
        lineNumber:
          type: integer
          minimum: 1
          description: Sequential line number
          example: 1
        type:
          $ref: '#/components/schemas/LineItemType'
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Item description
          example: "Web Development Services"
        quantity:
          type: number
          minimum: 0
          description: Quantity of items
          example: 40
        unitPrice:
          $ref: '#/components/schemas/Money'
        unitCost:
          $ref: '#/components/schemas/Money'
        taxRate:
          type: number
          minimum: 0
          maximum: 1
          description: Tax rate as decimal (0.15 = 15%)"
          example: 0.15
        discountType:
          $ref: '#/components/schemas/DiscountType'
        discountValue:
          type: number
          minimum: 0
          description: Discount value
          example: 0
        serviceCategoryId:
          type: string
          format: uuid
          description: Service category ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        rateCardId:
          type: string
          format: uuid
          description: Rate card ID for pricing
          example: "550e8400-e29b-41d4-a716-446655440001"
        metadata:
          type: object
          description: Additional metadata
          example: { "projectPhase": "development", "priority": "high" }
...
    CreateQuoteRequest:
      type: object
      required:
        - customerId
        - title
        - validFrom
        - validUntil
        - lineItems
      properties:
        customerId:
          type: string
          format: uuid
          description: Customer ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        projectId:
          type: string
          format: uuid
          description: Associated project ID
          example: "550e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          minLength: 3
          maxLength: 255
          description: Quote title
          example: "Website Development Project"
        description:
          type: string
          maxLength: 2000
          description: Quote description
          example: "Complete website development including design, development, and deployment"
        type:
          $ref: '#/components/schemas/QuoteType'
        validFrom:
          type: string
          format: date-time
          description: Quote validity start date
          example: "2025-01-01T00:00:00Z"
        validUntil:
          type: string
          format: date-time
          description: Quote validity end date
          example: "2025-12-31T23:59:59Z"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          default: "NZD"
          description: Quote currency
          example: "NZD"
        exchangeRate:
          type: number
          minimum: 0
          default: 1.0
          description: Exchange rate to base currency
          example: 1.0
        taxRate:
          type: number
          minimum: 0
          maximum: 1
          default: 0.15
          description: Quote-level tax rate
          example: 0.15
        discountType:
          $ref: '#/components/schemas/DiscountType'
        discountValue:
          type: number
          minimum: 0
          description: Quote-level discount value
          example: 0
        termsConditions:
          type: string
          maxLength: 5000
          description: Terms and conditions
          example: "Payment terms: 50% upfront, 50% on completion"
        notes:
          type: string
          maxLength: 2000
          description: Customer-facing notes
          example: "Includes 3 months of support after launch"
        internalNotes:
          type: string
          maxLength: 2000
          description: Internal notes (not visible to customer)
          example: "Customer prefers React over Vue"
        lineItems:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/LineItem'
...
    QuoteListResponse:
      type: object
      required:
        - quotes
        - pagination
      properties:
        quotes:
          type: array
          items:
            $ref: '#/components/schemas/Quote'
        pagination:
          type: object
          required:
            - page
            - pageSize
            - total
            - totalPages
          properties:
            page:
              type: integer
              minimum: 1
              description: Current page number
              example: 1
            pageSize:
              type: integer
              minimum: 1
              maximum: 100
              description: Number of items per page
              example: 20
            total:
              type: integer
              minimum: 0
              description: Total number of items
              example: 45
            totalPages:
              type: integer
              minimum: 0
              description: Total number of pages
              example: 3
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request data"
        code:
          type: string
          description: Error code for programmatic handling
          example: "INVALID_INPUT"
        details:
          type: object
          description: Additional error details
          example:
            field: "title"
            value: ""
            constraint: "minLength"
paths:
  /v1/quotes:
    get:
      summary: List quotes
      description: |
        Retrieve a paginated list of quotes for the current organization.
        
        ## Filtering
        - Filter by status, customer, project, type, or date range
        - Search by title or description using the `q` parameter
        - Filter by creator using `createdBy`
        
        ## Sorting
        - Sort by creation date, update date, title, status, total amount, or validity
        - Default sort is by creation date (newest first)
        
        ## Pagination
        - Page size can be 1-100 items (default: 20)
        - Page numbers start at 1
      tags:
        - Quotes
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, title, status, totalAmount, validUntil]
            default: createdAt
          description: Sort field
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/QuoteStatus'
          description: Filter by status
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by customer ID
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by project ID
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/QuoteType'
          description: Filter by quote type
        - name: q
          in: query
          schema:
            type: string
            maxLength: 100
          description: Search query for title or description
        - name: validFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by valid from date
        - name: validUntil
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by valid until date
        - name: createdBy
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by creator
      responses:
        '200':
          description: List of quotes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteListResponse'
              examples:
                success:
                  summary: Successful response
                  value:
                    quotes:
                      - id: "550e8400-e29b-41d4-a716-446655440000"
                        quoteNumber: "Q-2025-0001"
                        customerId: "550e8400-e29b-41d4-a716-446655440001"
                        title: "Website Development Project"
                        description: "Complete website development"
                        status: "draft"
                        type: "project"
                        validFrom: "2025-01-01T00:00:00Z"
                        validUntil: "2025-12-31T23:59:59Z"
                        currency: "NZD"
                        exchangeRate: 1.0
                        subtotal: "8000.00"
                        taxRate: 0.15
                        taxAmount: "1200.00"
                        discountType: "percentage"
                        discountValue: 0
                        discountAmount: "0.00"
                        totalAmount: "9200.00"
                        createdBy: "550e8400-e29b-41d4-a716-446655440003"
                        organizationId: "550e8400-e29b-41d4-a716-446655440005"
                        createdAt: "2025-01-01T08:00:00Z"
                        updatedAt: "2025-01-01T08:00:00Z"
                        lineItems: []
                    pagination:
                      page: 1
                      pageSize: 20
                      total: 1
                      totalPages: 1
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Tenant access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create quote
      description: |
        Create a new quote for a customer.
        
        ## Features
        - Automatic quote number generation with organization prefix
        - Automatic calculations (subtotal, tax, discounts, total)
        - Line item validation and calculation
        - Audit logging of quote creation
        - Multi-tenant isolation by organization
        
        ## Status Workflow
        New quotes are created with status `draft`. They can be transitioned through:
        `draft` → `pending` → `approved` → `sent` → `accepted`
        
        ## Calculations
        - Subtotal: Sum of all line item totals
        - Tax: Applied at quote level and line item level
        - Discounts: Can be percentage or fixed amount
        - Total: Subtotal + Tax - Discounts
      tags:
        - Quotes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
            examples:
              website_development:
                summary: Website Development Quote
                value:
                  customerId: "550e8400-e29b-41d4-a716-446655440001"
                  title: "Website Development Project"
                  description: "Complete website development including design, development, and deployment"
                  type: "project"
                  validFrom: "2025-01-01T00:00:00Z"
                  validUntil: "2025-12-31T23:59:59Z"
                  currency: "NZD"
                  exchangeRate: 1.0
                  taxRate: 0.15
                  discountType: "percentage"
                  discountValue: 0
                  termsConditions: "Payment terms: 50% upfront, 50% on completion"
                  notes: "Includes 3 months of support after launch"
                  internalNotes: "Customer prefers React over Vue"
                  lineItems:
                    - lineNumber: 1
                      type: "service"
                      description: "Web Development"
                      quantity: 40
                      unitPrice:
                        amount: "150.00"
                        currency: "NZD"
                      unitCost:
                        amount: "75.00"
                        currency: "NZD"
                      taxRate: 0.15
                      discountType: "percentage"
                      discountValue: 0
                      serviceCategoryId: "550e8400-e29b-41d4-a716-446655440000"
                      metadata:
                        projectPhase: "development"
                        priority: "high"
                    - lineNumber: 2
                      type: "service"
                      description: "Design Services"
                      quantity: 20
                      unitPrice:
                        amount: "100.00"
                        currency: "NZD"
                      unitCost:
                        amount: "50.00"
                        currency: "NZD"
                      taxRate: 0.15
                      discountType: "percentage"
                      discountValue: 0
                      serviceCategoryId: "550e8400-e29b-41d4-a716-446655440001"
                      metadata:
                        projectPhase: "design"
              consulting_service:
                summary: Consulting Service Quote
                value:
                  customerId: "550e8400-e29b-41d4-a716-446655440002"
                  title: "Business Strategy Consulting"
                  description: "Strategic business analysis and recommendations"
                  type: "service"
                  validFrom: "2025-01-01T00:00:00Z"
                  validUntil: "2025-03-31T23:59:59Z"
                  currency: "NZD"
                  exchangeRate: 1.0
                  taxRate: 0.15
                  discountType: "percentage"
                  discountValue: 10
                  termsConditions: "Payment within 30 days"
                  notes: "Includes follow-up consultation"
                  lineItems:
                    - lineNumber: 1
                      type: "service"
                      description: "Strategy Analysis"
                      quantity: 16
                      unitPrice:
                        amount: "200.00"
                        currency: "NZD"
                      unitCost:
                        amount: "100.00"
                        currency: "NZD"
                      taxRate: 0.15
                      discountType: "percentage"
                      discountValue: 0
                      serviceCategoryId: "550e8400-e29b-41d4-a716-446655440003"
      responses:
        '201':
          description: Quote created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
              examples:
                created_quote:
                  summary: Created quote response
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    quoteNumber: "Q-2025-0001"
                    customerId: "550e8400-e29b-41d4-a716-446655440001"
                    title: "Website Development Project"
                    description: "Complete website development including design, development, and deployment"
                    status: "draft"
                    type: "project"
                    validFrom: "2025-01-01T00:00:00Z"
                    validUntil: "2025-12-31T23:59:59Z"
                    currency: "NZD"
                    exchangeRate: 1.0
                    subtotal: "8000.00"
                    taxRate: 0.15
                    taxAmount: "1200.00"
                    discountType: "percentage"
                    discountValue: 0
                    discountAmount: "0.00"
                    totalAmount: "9200.00"
                    termsConditions: "Payment terms: 50% upfront, 50% on completion"
                    notes: "Includes 3 months of support after launch"
                    internalNotes: "Customer prefers React over Vue"
                    createdBy: "550e8400-e29b-41d4-a716-446655440003"
                    organizationId: "550e8400-e29b-41d4-a716-446655440005"
                    metadata:
                      source: "sales_team"
                      priority: "high"
                    createdAt: "2025-01-01T08:00:00Z"
                    updatedAt: "2025-01-01T08:00:00Z"
                    lineItems:
                      - id: "550e8400-e29b-41d4-a716-446655440006"
                        lineNumber: 1
                        type: "service"
                        description: "Web Development"
                        quantity: 40
                        unitPrice:
                          amount: "150.00"
                          currency: "NZD"
                        unitCost:
                          amount: "75.00"
                          currency: "NZD"
                        taxRate: 0.15
                        taxAmount: "900.00"
                        discountType: "percentage"
                        discountValue: 0
                        discountAmount: "0.00"
                        subtotal: "6000.00"
                        totalAmount: "6900.00"
                        serviceCategoryId: "550e8400-e29b-41d4-a716-446655440000"
                        metadata:
                          projectPhase: "development"
                          priority: "high"
                        createdAt: "2025-01-01T08:00:00Z"
                        updatedAt: "2025-01-01T08:00:00Z"
                      - id: "550e8400-e29b-41d4-a716-446655440007"
                        lineNumber: 2
                        type: "service"
                        description: "Design Services"
                        quantity: 20
                        unitPrice:
                          amount: "100.00"
                          currency: "NZD"
                        unitCost:
                          amount: "50.00"
                          currency: "NZD"
                        taxRate: 0.15
                        taxAmount: "300.00"
                        discountType: "percentage"
                        discountValue: 0
                        discountAmount: "0.00"
                        subtotal: "2000.00"
                        totalAmount: "2300.00"
                        serviceCategoryId: "550e8400-e29b-41d4-a716-446655440001"
                        metadata:
                          projectPhase: "design"
                        createdAt: "2025-01-01T08:00:00Z"
                        updatedAt: "2025-01-01T08:00:00Z"
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  summary: Validation error
                  value:
                    error: "ValidationError"
                    message: "Invalid request data"
                    code: "INVALID_INPUT"
                    details:
                      field: "title"
                      value: ""
                      constraint: "minLength"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Tenant access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...